.history/requirements_20260413192459.txt
# Core y Entorno
python-dotenv==1.1.0

# Telegram API
python-telegram-bot[job-queue]==22.7.0
python-telegram-bot-calendar==1.0.5

# Google API
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Testing
pytest==8.2.0
ruff==0.3.3
python-telegram-bot-calendar==1.0.5
pytest-asyncio==1.3.0

# Bases de datos
sqlalchemy==2.0.40
pymysql==1.1.1
cryptography==44.0.2
bcrypt==5.0.0

# API
fastapi==0.115.12
uvicorn[standard]==0.34.0

# HTTP requests (para pruebas)
requests==2.31.0
pymysql==1.1.1
cryptography==44.0.2

# API
fastapi==0.115.12
uvicorn[standard]>=0.34.0
bcrypt==5.0.0

# Servicios de Voz
gtts==2.5.1
